@@ -4,8 +4,8 @@ version: 1.2.0
 homepage: https://github.com/ferraridamiano/units_converter
 
 environment:
-  sdk: '>=2.14.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 dev_dependencies:
-  lints: ^1.0.1
+  lints: ^2.0.0
   test: ^1.21.1
