@@ -1,7 +1,7 @@
-import 'package:units_converter/models/node.dart';
+import 'package:units_converter/models/conversion_node.dart';
 import 'package:units_converter/models/property.dart';
 import 'package:units_converter/models/unit.dart';
-import 'package:units_converter/models/custom_conversion.dart';
+import 'package:units_converter/models/custom_property.dart';
 
 //Available ANGLE units
 enum ANGLE {
@@ -11,29 +11,7 @@ enum ANGLE {
   radians,
 }
 
-class Angle extends Property<ANGLE, double> {
-  /// Map between units and its symbol
-  static const Map<ANGLE, String?> mapSymbols = {
-    ANGLE.degree: '°',
-    ANGLE.minutes: "'",
-    ANGLE.seconds: "''",
-    ANGLE.radians: 'rad',
-  };
-
-  /// The number of significan figures to keep. E.g. 1.23456789) has 9
-  /// significant figures
-  int significantFigures;
-
-  /// Whether to remove the trailing zeros or not. E.g 1.00000000 has 9
-  /// significant figures and has trailing zeros. 1 has not trailing zeros.
-  bool removeTrailingZeros;
-
-  /// Whether to use the scientific notation (true) for [stringValue]s or
-  /// decimal notation (false)
-  bool useScientificNotation;
-
-  late CustomConversion _customConversion;
-
+class Angle extends CustomProperty {
   ///Class for angle conversions, e.g. if you want to convert 1 radiant in degree:
   ///```dart
   ///var angle = Angle(removeTrailingZeros: false);
@@ -41,44 +19,33 @@ class Angle extends Property<ANGLE, double> {
   ///print(ANGLE.degree);
   /// ```
   Angle(
-      {this.significantFigures = 10,
-      this.removeTrailingZeros = true,
-      this.useScientificNotation = true,
-      name}) {
-    this.name = name ?? PROPERTY.angle;
-    size = ANGLE.values.length;
-    Node conversionTree = Node(name: ANGLE.degree, leafNodes: [
-      Node(
-        coefficientProduct: 1 / 60,
-        name: ANGLE.minutes,
-      ),
-      Node(
-        coefficientProduct: 1 / 3600,
-        name: ANGLE.seconds,
-      ),
-      Node(
-        coefficientProduct: 57.295779513,
-        name: ANGLE.radians,
-      ),
-    ]);
-
-    _customConversion = CustomConversion(
-        conversionTree: conversionTree,
-        mapSymbols: mapSymbols,
-        significantFigures: significantFigures,
-        removeTrailingZeros: removeTrailingZeros,
-        useScientificNotation: useScientificNotation);
-  }
-
-  /// Converts a unit with a specific name (e.g. ANGLE.degree) and value to all
-  /// other units
-  @override
-  void convert(ANGLE name, double? value) =>
-      _customConversion.convert(name, value);
-  @override
-  List<Unit> getAll() => _customConversion.getAll();
-  @override
-  Unit getUnit(name) => _customConversion.getUnit(name);
+      {super.significantFigures,
+      super.removeTrailingZeros,
+      super.useScientificNotation,
+      name})
+      : super(
+          name: name ?? PROPERTY.angle,
+          mapSymbols: {
+            ANGLE.degree: '°',
+            ANGLE.minutes: "'",
+            ANGLE.seconds: "''",
+            ANGLE.radians: 'rad',
+          },
+          conversionTree: ConversionNode(name: ANGLE.degree, leafNodes: [
+            ConversionNode(
+              coefficientProduct: 1 / 60,
+              name: ANGLE.minutes,
+            ),
+            ConversionNode(
+              coefficientProduct: 1 / 3600,
+              name: ANGLE.seconds,
+            ),
+            ConversionNode(
+              coefficientProduct: 57.295779513,
+              name: ANGLE.radians,
+            ),
+          ]),
+        );
 
   Unit get degree => getUnit(ANGLE.degree);
   Unit get minutes => getUnit(ANGLE.minutes);
