@@ -1,7 +1,7 @@
-import 'package:units_converter/models/node.dart';
+import 'package:units_converter/models/conversion_node.dart';
 import 'package:units_converter/models/property.dart';
 import 'package:units_converter/models/unit.dart';
-import 'package:units_converter/models/custom_conversion.dart';
+import 'package:units_converter/models/custom_property.dart';
 
 //Available ENERGY units
 enum ENERGY {
@@ -13,31 +13,7 @@ enum ENERGY {
   energyFootPound,
 }
 
-class Energy extends Property<ENERGY, double> {
-  /// Map between units and its symbol
-  static const Map<ENERGY, String?> mapSymbols = {
-    ENERGY.joules: 'J',
-    ENERGY.calories: 'cal',
-    ENERGY.kilocalories: 'kcal',
-    ENERGY.kilowattHours: 'kwh',
-    ENERGY.electronvolts: 'eV',
-    ENERGY.energyFootPound: 'ft⋅lbf',
-  };
-
-  /// The number of significan figures to keep. E.g. 1.23456789) has 9
-  /// significant figures
-  int significantFigures;
-
-  /// Whether to remove the trailing zeros or not. E.g 1.00000000 has 9
-  /// significant figures and has trailing zeros. 1 has not trailing zeros.
-  bool removeTrailingZeros;
-
-  /// Whether to use the scientific notation (true) for [stringValue]s or
-  /// decimal notation (false)
-  bool useScientificNotation;
-
-  late CustomConversion _customConversion;
-
+class Energy extends CustomProperty {
   ///Class for energy conversions, e.g. if you want to convert 1 joule in kilowatt hours:
   ///```dart
   ///var energy = Energy(removeTrailingZeros: false);
@@ -45,53 +21,44 @@ class Energy extends Property<ENERGY, double> {
   ///print(ENERGY.kilowatt_hours);
   /// ```
   Energy(
-      {this.significantFigures = 10,
-      this.removeTrailingZeros = true,
-      this.useScientificNotation = true,
-      name}) {
-    this.name = name ?? PROPERTY.energy;
-    size = ENERGY.values.length;
-    Node conversionTree = Node(name: ENERGY.joules, leafNodes: [
-      Node(
-        coefficientProduct: 4.1867999409,
-        name: ENERGY.calories,
-        leafNodes: [
-          Node(
-            coefficientProduct: 1000.0,
-            name: ENERGY.kilocalories,
-          ),
-        ],
-      ),
-      Node(
-        coefficientProduct: 3600000.0,
-        name: ENERGY.kilowattHours,
-      ),
-      Node(
-        coefficientProduct: 1.60217646e-19,
-        name: ENERGY.electronvolts,
-      ),
-      Node(
-        coefficientProduct: 1 / 1.3558179483314004,
-        name: ENERGY.energyFootPound,
-      ),
-    ]);
-
-    _customConversion = CustomConversion(
-        conversionTree: conversionTree,
-        mapSymbols: mapSymbols,
-        significantFigures: significantFigures,
-        removeTrailingZeros: removeTrailingZeros,
-        useScientificNotation: useScientificNotation);
-  }
-
-  ///Converts a unit with a specific name (e.g. ENERGY.calories) and value to all other units
-  @override
-  void convert(ENERGY name, double? value) =>
-      _customConversion.convert(name, value);
-  @override
-  List<Unit> getAll() => _customConversion.getAll();
-  @override
-  Unit getUnit(name) => _customConversion.getUnit(name);
+      {super.significantFigures,
+      super.removeTrailingZeros,
+      super.useScientificNotation,
+      name})
+      : super(
+            name: name ?? PROPERTY.energy,
+            mapSymbols: {
+              ENERGY.joules: 'J',
+              ENERGY.calories: 'cal',
+              ENERGY.kilocalories: 'kcal',
+              ENERGY.kilowattHours: 'kwh',
+              ENERGY.electronvolts: 'eV',
+              ENERGY.energyFootPound: 'ft⋅lbf',
+            },
+            conversionTree: ConversionNode(name: ENERGY.joules, leafNodes: [
+              ConversionNode(
+                coefficientProduct: 4.1867999409,
+                name: ENERGY.calories,
+                leafNodes: [
+                  ConversionNode(
+                    coefficientProduct: 1000.0,
+                    name: ENERGY.kilocalories,
+                  ),
+                ],
+              ),
+              ConversionNode(
+                coefficientProduct: 3600000.0,
+                name: ENERGY.kilowattHours,
+              ),
+              ConversionNode(
+                coefficientProduct: 1.60217646e-19,
+                name: ENERGY.electronvolts,
+              ),
+              ConversionNode(
+                coefficientProduct: 1 / 1.3558179483314004,
+                name: ENERGY.energyFootPound,
+              ),
+            ]));
 
   Unit get joules => getUnit(ENERGY.joules);
   Unit get calories => getUnit(ENERGY.calories);
