@@ -1,7 +1,7 @@
-import 'package:units_converter/models/node.dart';
+import 'package:units_converter/models/conversion_node.dart';
 import 'package:units_converter/models/property.dart';
 import 'package:units_converter/models/unit.dart';
-import 'package:units_converter/models/custom_conversion.dart';
+import 'package:units_converter/models/custom_property.dart';
 
 //Available FORCE units
 enum FORCE {
@@ -12,30 +12,7 @@ enum FORCE {
   poundal,
 }
 
-class Force extends Property<FORCE, double> {
-  /// Map between units and its symbol
-  static const Map<FORCE, String?> mapSymbols = {
-    FORCE.newton: 'N',
-    FORCE.dyne: 'dyn',
-    FORCE.poundForce: 'lbf',
-    FORCE.kilogramForce: 'kgf',
-    FORCE.poundal: 'pdl',
-  };
-
-  /// The number of significan figures to keep. E.g. 1.23456789) has 9
-  /// significant figures
-  int significantFigures;
-
-  /// Whether to remove the trailing zeros or not. E.g 1.00000000 has 9
-  /// significant figures and has trailing zeros. 1 has not trailing zeros.
-  bool removeTrailingZeros;
-
-  /// Whether to use the scientific notation (true) for [stringValue]s or
-  /// decimal notation (false)
-  bool useScientificNotation;
-
-  late CustomConversion _customConversion;
-
+class Force extends CustomProperty {
   ///Class for force conversions, e.g. if you want to convert 1 newton in pound force:
   ///```dart
   ///var force = Force(removeTrailingZeros: false);
@@ -43,47 +20,37 @@ class Force extends Property<FORCE, double> {
   ///print(FORCE.pound_force);
   /// ```
   Force(
-      {this.significantFigures = 10,
-      this.removeTrailingZeros = true,
-      this.useScientificNotation = true,
-      name}) {
-    this.name = name ?? PROPERTY.force;
-    size = FORCE.values.length;
-    Node conversionTree = Node(name: FORCE.newton, leafNodes: [
-      Node(
-        coefficientProduct: 1e-5,
-        name: FORCE.dyne,
-      ),
-      Node(
-        coefficientProduct: 4.4482216152605,
-        name: FORCE.poundForce,
-      ),
-      Node(
-        coefficientProduct: 9.80665,
-        name: FORCE.kilogramForce,
-      ),
-      Node(
-        coefficientProduct: 0.138254954376,
-        name: FORCE.poundal,
-      ),
-    ]);
-
-    _customConversion = CustomConversion(
-        conversionTree: conversionTree,
-        mapSymbols: mapSymbols,
-        significantFigures: significantFigures,
-        removeTrailingZeros: removeTrailingZeros,
-        useScientificNotation: useScientificNotation);
-  }
-
-  ///Converts a unit with a specific name (e.g. FORCE.newton) and value to all other units
-  @override
-  void convert(FORCE name, double? value) =>
-      _customConversion.convert(name, value);
-  @override
-  List<Unit> getAll() => _customConversion.getAll();
-  @override
-  Unit getUnit(name) => _customConversion.getUnit(name);
+      {super.significantFigures,
+      super.removeTrailingZeros,
+      super.useScientificNotation,
+      name})
+      : super(
+            name: name ?? PROPERTY.force,
+            mapSymbols: {
+              FORCE.newton: 'N',
+              FORCE.dyne: 'dyn',
+              FORCE.poundForce: 'lbf',
+              FORCE.kilogramForce: 'kgf',
+              FORCE.poundal: 'pdl',
+            },
+            conversionTree: ConversionNode(name: FORCE.newton, leafNodes: [
+              ConversionNode(
+                coefficientProduct: 1e-5,
+                name: FORCE.dyne,
+              ),
+              ConversionNode(
+                coefficientProduct: 4.4482216152605,
+                name: FORCE.poundForce,
+              ),
+              ConversionNode(
+                coefficientProduct: 9.80665,
+                name: FORCE.kilogramForce,
+              ),
+              ConversionNode(
+                coefficientProduct: 0.138254954376,
+                name: FORCE.poundal,
+              ),
+            ]));
 
   Unit get newton => getUnit(FORCE.newton);
   Unit get dyne => getUnit(FORCE.dyne);
