@@ -1,7 +1,7 @@
-import 'package:units_converter/models/node.dart';
+import 'package:units_converter/models/conversion_node.dart';
 import 'package:units_converter/models/property.dart';
 import 'package:units_converter/models/unit.dart';
-import 'package:units_converter/models/custom_conversion.dart';
+import 'package:units_converter/models/custom_property.dart';
 
 //Available PRESSURE units
 enum PRESSURE {
@@ -15,33 +15,7 @@ enum PRESSURE {
   inchOfMercury,
 }
 
-class Pressure extends Property<PRESSURE, double> {
-  /// Map between units and its symbol
-  static const Map<PRESSURE, String?> mapSymbols = {
-    PRESSURE.pascal: 'Pa',
-    PRESSURE.atmosphere: 'atm',
-    PRESSURE.bar: 'bar',
-    PRESSURE.millibar: 'mbar',
-    PRESSURE.psi: 'psi',
-    PRESSURE.torr: 'torr', //Same as mmHg
-    PRESSURE.hectoPascal: 'hPa',
-    PRESSURE.inchOfMercury: 'inHg',
-  };
-
-  /// The number of significan figures to keep. E.g. 1.23456789) has 9
-  /// significant figures
-  int significantFigures;
-
-  /// Whether to remove the trailing zeros or not. E.g 1.00000000 has 9
-  /// significant figures and has trailing zeros. 1 has not trailing zeros.
-  bool removeTrailingZeros;
-
-  /// Whether to use the scientific notation (true) for [stringValue]s or
-  /// decimal notation (false)
-  bool useScientificNotation;
-
-  late CustomConversion _customConversion;
-
+class Pressure extends CustomProperty {
   ///Class for pressure conversions, e.g. if you want to convert 1 bar in atmosphere:
   ///```dart
   ///var pressure = Pressure(removeTrailingZeros: false);
@@ -49,53 +23,56 @@ class Pressure extends Property<PRESSURE, double> {
   ///print(PRESSURE.atmosphere);
   /// ```
   Pressure(
-      {this.significantFigures = 10,
-      this.removeTrailingZeros = true,
-      this.useScientificNotation = true,
-      name}) {
-    this.name = name ?? PROPERTY.pressure;
-    size = PRESSURE.values.length;
-    Node conversionTree = Node(name: PRESSURE.pascal, leafNodes: [
-      Node(coefficientProduct: 101325.0, name: PRESSURE.atmosphere, leafNodes: [
-        Node(coefficientProduct: 0.987, name: PRESSURE.bar, leafNodes: [
-          Node(
-            coefficientProduct: 1e-3,
-            name: PRESSURE.millibar,
-          ),
-        ]),
-      ]),
-      Node(
-        coefficientProduct: 6894.757293168,
-        name: PRESSURE.psi,
-      ),
-      Node(coefficientProduct: 133.322368421, name: PRESSURE.torr, leafNodes: [
-        Node(
-          coefficientProduct: 25.4,
-          name: PRESSURE.inchOfMercury,
-        )
-      ]),
-      Node(
-        coefficientProduct: 1e2,
-        name: PRESSURE.hectoPascal,
-      )
-    ]);
-
-    _customConversion = CustomConversion(
-        conversionTree: conversionTree,
-        mapSymbols: mapSymbols,
-        significantFigures: significantFigures,
-        removeTrailingZeros: removeTrailingZeros,
-        useScientificNotation: useScientificNotation);
-  }
-
-  ///Converts a unit with a specific name (e.g. PRESSURE.psi) and value to all other units
-  @override
-  void convert(PRESSURE name, double? value) =>
-      _customConversion.convert(name, value);
-  @override
-  List<Unit> getAll() => _customConversion.getAll();
-  @override
-  Unit getUnit(name) => _customConversion.getUnit(name);
+      {super.significantFigures,
+      super.removeTrailingZeros,
+      super.useScientificNotation,
+      name})
+      : super(
+          name: name ?? PROPERTY.pressure,
+          mapSymbols: {
+            PRESSURE.pascal: 'Pa',
+            PRESSURE.atmosphere: 'atm',
+            PRESSURE.bar: 'bar',
+            PRESSURE.millibar: 'mbar',
+            PRESSURE.psi: 'psi',
+            PRESSURE.torr: 'torr', //Same as mmHg
+            PRESSURE.hectoPascal: 'hPa',
+            PRESSURE.inchOfMercury: 'inHg',
+          },
+          conversionTree: ConversionNode(name: PRESSURE.pascal, leafNodes: [
+            ConversionNode(
+                coefficientProduct: 101325.0,
+                name: PRESSURE.atmosphere,
+                leafNodes: [
+                  ConversionNode(
+                      coefficientProduct: 0.987,
+                      name: PRESSURE.bar,
+                      leafNodes: [
+                        ConversionNode(
+                          coefficientProduct: 1e-3,
+                          name: PRESSURE.millibar,
+                        ),
+                      ]),
+                ]),
+            ConversionNode(
+              coefficientProduct: 6894.757293168,
+              name: PRESSURE.psi,
+            ),
+            ConversionNode(
+                coefficientProduct: 133.322368421,
+                name: PRESSURE.torr,
+                leafNodes: [
+                  ConversionNode(
+                    coefficientProduct: 25.4,
+                    name: PRESSURE.inchOfMercury,
+                  )
+                ]),
+            ConversionNode(
+              coefficientProduct: 1e2,
+              name: PRESSURE.hectoPascal,
+            )
+          ]),
+        );
 
   Unit get pascal => getUnit(PRESSURE.pascal);
   Unit get atmosphere => getUnit(PRESSURE.atmosphere);
