@@ -1,7 +1,7 @@
-import 'package:units_converter/models/node.dart';
+import 'package:units_converter/models/conversion_node.dart';
 import 'package:units_converter/models/property.dart';
 import 'package:units_converter/models/unit.dart';
-import 'package:units_converter/models/custom_conversion.dart';
+import 'package:units_converter/models/custom_property.dart';
 
 //Available TORQUE units
 enum TORQUE {
@@ -12,30 +12,7 @@ enum TORQUE {
   poundalMeter,
 }
 
-class Torque extends Property<TORQUE, double> {
-  /// Map between units and its symbol
-  static const Map<TORQUE, String?> mapSymbols = {
-    TORQUE.newtonMeter: 'N·m',
-    TORQUE.dyneMeter: 'dyn·m',
-    TORQUE.poundForceFeet: 'lbf·ft',
-    TORQUE.kilogramForceMeter: 'kgf·m',
-    TORQUE.poundalMeter: 'pdl·m',
-  };
-
-  /// The number of significan figures to keep. E.g. 1.23456789) has 9
-  /// significant figures
-  int significantFigures;
-
-  /// Whether to remove the trailing zeros or not. E.g 1.00000000 has 9
-  /// significant figures and has trailing zeros. 1 has not trailing zeros.
-  bool removeTrailingZeros;
-
-  /// Whether to use the scientific notation (true) for [stringValue]s or
-  /// decimal notation (false)
-  bool useScientificNotation;
-
-  late CustomConversion _customConversion;
-
+class Torque extends CustomProperty {
   ///Class for torque conversions, e.g. if you want to convert 1 square meters in acres:
   ///```dart
   ///var torque = Torque(removeTrailingZeros: false);
@@ -43,47 +20,38 @@ class Torque extends Property<TORQUE, double> {
   ///print(TORQUE.acres);
   /// ```
   Torque(
-      {this.significantFigures = 10,
-      this.removeTrailingZeros = true,
-      this.useScientificNotation = true,
-      name}) {
-    this.name = name ?? PROPERTY.torque;
-    size = TORQUE.values.length;
-    Node conversionTree = Node(name: TORQUE.newtonMeter, leafNodes: [
-      Node(
-        coefficientProduct: 1e-5,
-        name: TORQUE.dyneMeter,
-      ),
-      Node(
-        coefficientProduct: 1.35581794902490555,
-        name: TORQUE.poundForceFeet,
-      ),
-      Node(
-        coefficientProduct: 9.807,
-        name: TORQUE.kilogramForceMeter,
-      ),
-      Node(
-        coefficientProduct: 0.138254954376,
-        name: TORQUE.poundalMeter,
-      ),
-    ]);
-
-    _customConversion = CustomConversion(
-        conversionTree: conversionTree,
-        mapSymbols: mapSymbols,
-        significantFigures: significantFigures,
-        removeTrailingZeros: removeTrailingZeros,
-        useScientificNotation: useScientificNotation);
-  }
-
-  ///Converts a unit with a specific name (e.g. TORQUE.newton_meter) and value to all other units
-  @override
-  void convert(TORQUE name, double? value) =>
-      _customConversion.convert(name, value);
-  @override
-  List<Unit> getAll() => _customConversion.getAll();
-  @override
-  Unit getUnit(name) => _customConversion.getUnit(name);
+      {super.significantFigures,
+      super.removeTrailingZeros,
+      super.useScientificNotation,
+      name})
+      : super(
+          name: name ?? PROPERTY.torque,
+          mapSymbols: {
+            TORQUE.newtonMeter: 'N·m',
+            TORQUE.dyneMeter: 'dyn·m',
+            TORQUE.poundForceFeet: 'lbf·ft',
+            TORQUE.kilogramForceMeter: 'kgf·m',
+            TORQUE.poundalMeter: 'pdl·m',
+          },
+          conversionTree: ConversionNode(name: TORQUE.newtonMeter, leafNodes: [
+            ConversionNode(
+              coefficientProduct: 1e-5,
+              name: TORQUE.dyneMeter,
+            ),
+            ConversionNode(
+              coefficientProduct: 1.35581794902490555,
+              name: TORQUE.poundForceFeet,
+            ),
+            ConversionNode(
+              coefficientProduct: 9.807,
+              name: TORQUE.kilogramForceMeter,
+            ),
+            ConversionNode(
+              coefficientProduct: 0.138254954376,
+              name: TORQUE.poundalMeter,
+            ),
+          ]),
+        );
 
   Unit get newtonMeter => getUnit(TORQUE.newtonMeter);
   Unit get dyneMeter => getUnit(TORQUE.dyneMeter);
